Adds ROS messages for Beckhoff devices EL3162 and EL4102.
Summary:
Adds ROS messages for Beckhoff devices EL3162 and EL4102. Device
EL3162 is a 2-channel analog input terminal for voltage range 0-10 V.
Device EL4102 is a 2-channel analog output terminal of voltage range
0-10 V.

Test Plan:
Tested together with fcat. Verified that the state ROS messages are
published properly when the devices are present in the EtherCAT
network, and made sure that the commands sent to the EL4102 via the
newly added ROS messages are executed as expected.

Reviewers: JosephBowkett, alex-brinkman